delete old two-way diff stuff and consolidate modules again
